1. changed drone action grid controls to appear and disappear through visibility. 2. changed drone id to int? 3. icon sorted for all windows

+++ BL/BL_Actions.cs
@@ -201,7 +201,7 @@ namespace BL
                 Scheduled = prc.Linked,
                 PickedUp = prc.PickedUp,
                 Delivered = DateTime.Now,
-                DroneId = dr.Id,
+                DroneId = (int)dr.Id,
             });
         }
     }
